refactors and splitting into files
